requirements.txt
Flask==3.0
bcrypt==4.0.1
gunicorn==22.0
google-generativeai==0.3.2
python-dotenv==1.0.0